test: ProductListViewModel 테스트 작성
